There are no more DirectoryCopyTask. Now we only have copy file tasks (and the purge tasks).
Rationale :

Support to backup the repertory /home/daniel/video that is not yet
present at all in the backup directory.

If that directory contains subdirectories
 /home/daniel/Video/foo
 /home/daniel/Video/bar
If /home/daniel/Video/foo is excluded, it was copied because the task
was to copy the whole /home/daniel/video regardless of its internal
structure.

By the way, the non backuped directories are created in the backup
directory but left empty.