Make the terminal command line available everywhere.
'CommandLine' needs to know the command line which launches the terminal
(e.g. "terminology") and which launches a command in a terminal (e.g. "konsole -e").
These commands are written in the configuration file 'lora.cfg' and thus
available in the object 'Configuration'.
A 'Configuration*' pointer is propagated more or less everywhere.