Fixing a bug in the purge process.
Let say we removed the directory ~/foo

Then it creates the task to move <backup>/foo to <purge>/foo,
but in the same time the purge loop enters in  <backup>/foo and, for example,
found the directory <backup>/foo/bar. The task of moving
<backup>/foo/bar is created.

Of course this second task will crash because when the second running,
the first is already done and <backup>/foo does not exists anymore.