Add an exception : if a file is in a directory containing "/.git/object", we do the backup even if the time of the backup is more recent than the original one (situation that should not happen)
Here is more or less the reason :

It seems that Git is dealing with the timestamp of its object files
in a strange way.
- A clone seem to create files with date of the clone time
- When pulling back, some object files retrieve their original
  time.

=> The timestamp criterion often failed on the cloned directories.

When time_backup > time original for a git object file, and when the
sizes are equal, we backup the file and add a line in the log file.